refactor common functions to utils.py
